refactor: processor 가독성 수정
